% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BarGene.R
\name{barGene}
\alias{barGene}
\title{Create a bar plot of gene expression}
\usage{
barGene(
  genes,
  counts,
  conditions,
  title = "Gene expression",
  norm = NULL,
  eb = "sd",
  returnDat = F,
  col = "Dark2"
)
}
\arguments{
\item{genes}{character vector of genes to plot. Must match rownames(counts)}

\item{counts}{count matrix of gene expression, where rows are genes and columns are samples}

\item{conditions}{character vector of length ncol(counts) describing the condition of each sampmle in counts}

\item{title}{character describing the title of the plot}

\item{norm}{character describing the condition to use as a reference for relative gene expression. Each gene's expression will be set relative to this condition. Leave 'NULL' if plotting raw values.}

\item{eb}{character describing the style of error bar. "sd" = standard deviation, "se" = standard error of the mean. Use '0' if no error bars to be plotted. Default is 'sd'.}

\item{returnDat}{Boolean indicating if list of raw and summarized data should be returned for further analysis. Default is FALSE.}

\item{col}{character indicating the RColorBrewer palette name to be used. Default is "Dark2"}
}
\value{
Bar plot of gene expression and list of length 2 containing 'rawData' and 'Summary' of gene expression data if 'returnDat' is TRUE.
}
\description{
This function will accept a list of genes and counts to produce a bar plot of gene expression in different conditions
}
