% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barplotTopGSEA.R
\name{barplotTopGSEA}
\alias{barplotTopGSEA}
\title{Plot top n pathways according to given stat}
\usage{
barplotTopGSEA(
  fgRes,
  n = 30,
  by = "padj",
  decreasing = FALSE,
  parseBader = T,
  plot.title = "GSEA Bar plot"
)
}
\arguments{
\item{n}{number of pathways to plot, default is 30.}

\item{by}{The stat for selecting top pathways, default is padj}

\item{decreasing}{Should the sort order be increasing or decreasing? Default is False.}

\item{parseBader}{Boolean indicating if gene set names should be parsed for the output figure following Bader Lab nomenclature (i.e. '\%' as delimiter)}

\item{plot.title}{arbitrary plot title}

\item{fgsea}{GSEA results from GSEA() function}
}
\description{
Plot top n pathways according to given stat
}
