% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcGaps.R
\name{calcGaps}
\alias{calcGaps}
\title{Calculate optimal number of clusters}
\usage{
calcGaps(resList, maxClus = 10, iter = 10)
}
\arguments{
\item{resList}{List of results data frames}

\item{maxClus}{Maximum number of clusters to test, minimum 3. Default 10. NOTE: The more, the longer it takes.}

\item{iter}{Number of bootstrapping iterations for calculating gap statistic. Default 10. NOTE: the more, the longer it takes.}
}
\value{
A vector of length 2 containing the maximum gap statistic and the optimal number of k-means clusters, respectively.
}
\description{
Calculate the gap statistic for a series of k-means clusterings using the SAGx
package. Use the number of clusters with the highest gap statistic for your
clustering analysis.
}
