% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusFigs.R
\name{clusFigs}
\alias{clusFigs}
\title{k-means clustered figures}
\usage{
clusFigs(resList, numClus, title = "Clustered Results", col = "Dark2")
}
\arguments{
\item{resList}{A list of results data frames. names(resList) will be used}

\item{numClus}{Number of k-means clusters to cluster the results}

\item{title}{Character indicating the titles of the plots to be made}

\item{col}{Character indicating the RColorBrewer palette name or list of colours (hex, name, rgb()) to be used for the bar plot. Default is "Dark2"}
}
\value{
A data frame of log2FoldChanges for each comparison as columns (rows are genes) and a column named "cluster" indicating the cluster each gene belongs to. Two figures: a Heatmap of log2FoldChange of each gene ordered into clusters and a bar plot of the average log2FoldChange in each cluster (+/- SD) by comparison.
}
\description{
This function is for use with more than one results object. Combine the results
from multiple comparisons by clustering genes based on their log2 fold changes
between conditions for each comparison. This function will create a clustered
heatmap and bar plot of the average log2 fold-change in each cluster.
}
