% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusGO.R
\name{clusGO}
\alias{clusGO}
\title{Pathway Enrichment Analysis on Clusters}
\usage{
clusGO(
  clusRes,
  species = "hsapiens",
  bg = rownames(clusRes),
  source = NULL,
  corr = "fdr",
  iea = FALSE,
  prefix = "ClusGO",
  ts = c(10, 500),
  pdf = TRUE,
  fig = TRUE,
  returnGost = FALSE,
  writeRes = TRUE,
  writeGem = FALSE,
  returnRes = FALSE,
  writeGene = FALSE
)
}
\arguments{
\item{clusRes}{data.frame object that is output from clusFigs()}

\item{writeGene}{Boolean indicating if the genes in each cluster should be written out to '\emph{prefix}_cluster\emph{n}.genes.txt". Default is FALSE.}

\item{...}{Arguments for GO_GEM() see ?GO_GEM for details. \emph{bg} is automatically set as rownames(clusRes).}
}
\value{
For each cluster, output from GO_GEM() is generated and output to "\emph{prefix}\emph{cluster}\emph{n}"
}
\description{
Run Pathway Enrichment Analysis (same as GO_GEM()) on output from clusFigs()
}
