% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusFigs.R
\name{clusRelev}
\alias{clusRelev}
\title{Reorder clusters generated from clusFigs() and replot.}
\usage{
clusRelev(
  clusRes,
  cluslev,
  rename = T,
  title = "Releveled Clusters",
  col = "Dark2",
  hmcol = NULL,
  labgenes = "",
  avgExp = F,
  showStat = T,
  retStat = F
)
}
\arguments{
\item{clusRes}{Results dataframe from clusRes()}

\item{cluslev}{Numeric vector indicating the new order of clusters for clusRes}

\item{rename}{Boolean indicating if clusteres should be renamed based on their new order. Default=T}

\item{title}{Character indicating the title for the plots. Default= "Releveled Clsuters"}

\item{col}{Character indicating the RColorBrewer palette name or list of colours (hex, name, rgb()) to be used for the bar plot. Default is "Dark2"}

\item{hmcol}{Colour Ramp Palette of length 100 indicating the colour palette of the heatmap. Leave NULL for default.}

\item{labgenes}{Character vector corresponding to rownames in clusRes of genes to label in heatmap. Set to NULL to label all genes. Default is "" which labeles no genes.}

\item{avgExp}{Boolean. Set to TRUE if clusRes was generated with 'avgExp=T' in clusFigs().}

\item{showStat}{Boolean. If set to TRUE, significance symbols will be shown for significant contrasts in boxplot.}

\item{retStat}{Boolean. If set to TRUE, stats for contrasts (p-values and FDR) will be returned for each cluster along with results in a list object.}
}
\value{
A data frame of log2FoldChanges for each comparison as columns (rows are genes) and a column named "cluster" indicating the cluster each gene belongs to. Two figures: a Heatmap of log2FoldChange of each gene ordered into clusters and a bar plot of the average log2FoldChange in each cluster (+/- SD) by comparison.
}
\description{
Reorder clusters generated from clusFigs() and replot.
}
