% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colPal.R
\name{colPal}
\alias{colPal}
\title{Colour palettes}
\usage{
colPal(col)
}
\arguments{
\item{col}{A character vector indicating the custom colours ("rgb", "colour name", "Hexadecimal") or RColourBrewer palette to be used. If list of colour values, the must all be of the same format ("color name", "hexadecimal", or "rgb()") If RColourBrewer palette name, length must be 1.}
}
\value{
A character vector of hexadecimal values representing the colour palette
}
\description{
This function creates either a custom colour palette or can return the colours
from an RColorBrewer palette
}
