% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadData.R
\name{combCounts}
\alias{combCounts}
\title{Combine two count tables}
\usage{
combCounts(x, y, xName = NULL, yName = NULL)
}
\arguments{
\item{x}{data.frame object representing the first count table. Entries must be numeric with columns representing samples and rownames are unique gene names.}

\item{y}{data.frame object representing the second count table following the same rules as x. Both x and y must have shared rownames (gene) meaning the nomenclature must be the same and there must be at least SOME shared genes. Genes not shared between samples will be removed. Gene order does not matter.}

\item{xName}{Character indicating an optional identifier added to the colnames of x. Default is NULL.}

\item{yName}{Character indicating an optional identifier added to the colnames of y. Default is NULL.}
}
\value{
data.frame object of combined x and y where the rownames are shared genes and columns are samples.
}
\description{
Combine two count tables
}
