% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GO.R
\name{combGO_plot}
\alias{combGO_plot}
\title{Combine GO Results from DE analysis into one plot}
\usage{
combGO_plot(
  GOresList,
  title = "GO results",
  ts = c(10, 500),
  sig = T,
  numTerm = 10,
  upcols = c("lightpink", "red"),
  downcols = c("lightblue", "blue"),
  labs = c("Downregulated", "Upregulated")
)
}
\arguments{
\item{GOresList}{list of GO_GEM results (set returnRes=T) of length 2. Best for DEG GO results, where the first entry in the list is downregulated, and the second entry in the list is upregulated}

\item{title}{Character for the plot title}

\item{ts}{numeric vector of length 2 with the minimum and maximum term sizes for plotting. Default is c(10,500)}

\item{sig}{Boolean indicating if top significant results should be plotted. Set to FALSE to plot top enriched results}

\item{numTerm}{Numeric indicating the number of top terms to plot. Default is 10.}

\item{upcols}{character vector of length 2 indicating the colour for upregulated enrichment and significance bars}

\item{downcols}{character vector of length 2 indicating the colour for downregulated enrichment and significance bars}

\item{labs}{character vector of length 2 indicating the specific legend labels for the entries in GOresList. Default=c("Downregulated","Upgreulated")}
}
\value{
A plot of top GO results
}
\description{
This function will combine GO results from GO_GEM when performed on DEGs into a single plot.
}
