% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Correlations.R
\name{corHeat}
\alias{corHeat}
\title{Sample correlation heatmap}
\usage{
corHeat(
  counts,
  method = "spearman",
  title = "Spearman Correlation",
  hmcol = NULL,
  showNum = TRUE,
  showTree = FALSE,
  annodf = NULL,
  annoCols = "Dark2"
)
}
\arguments{
\item{counts}{data.frame object with rows=genes, and columns=samples}

\item{method}{character indicating correlation method, default is "spearman"}

\item{title}{character indicating plot title}

\item{hmcol}{Color ramp palette to use for heatmap colours. Defaults to 'RdYlBu' colours}

\item{showNum}{Boolean indicating if correlation coefficients should be displayed on heatmap. Default=TRUE.}

\item{showTree}{Boolean indicating if hierarchical clustering trees should be shown. Default=FALSE}

\item{annodf}{Data frame with rownames=sample names and columns of annotation data to be plotted. Leave NULL if none.}

\item{annoCols}{Colour palette for annotations}
}
\value{
Heatmap with clustered rows/columns of correlation between samples
}
\description{
Sample correlation heatmap
}
