% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_plot.R
\name{count_plot}
\alias{count_plot}
\title{A function to make a violin plot of normalized counts}
\usage{
count_plot(
  counts,
  scaling = "zscore",
  genes,
  condition,
  title = "expression",
  compare = NULL
)
}
\arguments{
\item{counts}{Normalized counts from a DESeq2 object - use 'counts(dds, normalized=T)'}

\item{scaling}{Method used to scale counts per gene across samples. Either 'zscore' or 'log10'. Default is 'zscore'}

\item{genes}{Character vector of genes to subset from counts. Must correspond with rownames(counts).}

\item{condition}{Character vector of conditions in DESeq2 object. Must be in order of columns (counts).}

\item{title}{Character vector indicating title of plot. Defaults to "expression"}

\item{compare}{List of character vectors (each of length 2) indicating pairwise comparisons. If NULL, all possible comparisons will be made. Default is NULL}
}
\value{
Generates a violin plot
}
\description{
This function takes normalized counts of specific genes from a DESeq2 counts
object, scales them, and creates a violin plot with pairwise t-tests by condition
}
