% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmt.R
\name{customGMT}
\alias{customGMT}
\title{Make custom gene sets}
\usage{
customGMT(gost, key, gmt)
}
\arguments{
\item{gost}{The gost output of GO_GEM (get by setting returnGost=T when using GO_GEM)}

\item{key}{The keyword used to identify GO terms of interest}

\item{gmt}{The geneset object (read in using qusage::read.gmt())}
}
\value{
A list of gene sets in the format for use with other BinfTools functions
}
\description{
This function will use the gost output of GO_GEM along with a complete geneset
object (downloaded from gprofiler) to extract gene sets of GO terms containing
a given key.
}
