% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GO.R
\name{customGO}
\alias{customGO}
\title{Custom GO Function}
\usage{
customGO(
  genes,
  gmt,
  gsName = "custom GeneSet",
  bg = NULL,
  sp = "human",
  FDR = T,
  byRegion = F,
  enr = "pos",
  significant = T,
  minp = 1e-300
)
}
\arguments{
\item{genes}{character vector of genes to be queried}

\item{gmt}{Named list of gene set(s) to be analyzed for enrichment}

\item{gsName}{Character specifying geneSet source - to be incorportated into output and help indicate source of gene sets}

\item{bg}{character vector of genes indicating the background of genes. If left NULL, argument 'sp' will indicate species and assume all genes are in background.}

\item{sp}{character of either "human" or "mouse" indicating the species. This will provide the number of background genes if bg=NULL.}

\item{FDR}{Boolean indicating if p-values should be FDR corrected (Benjamini-Hochberg). Default is TRUE.}

\item{byRegion}{Boolean. If query genes has duplicate gene symbols, set to TRUE to remove duplicates (relevant only if query genes are from annotated peaks from genomic data). Default=F.}

\item{enr}{character of either "pos" or "neg" indicating if significance should be calculated for positive or negative hypergeometric enrichment, respectively. default="pos".}

\item{significant}{Boolean indicating if only significant (p<0.05) results should be returned. defaulte=T.}

\item{minp}{numeric indicating the minimum p-value possible reported (to avoid zeros)}
}
\value{
Data.frame of same structure of results table from GO_GEM() when returnRes=T.
}
\description{
Perform a hypergeometric enrichment test for a custom gene set
}
