% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/densityPlot.R
\name{dPlot}
\alias{dPlot}
\title{Density Plot}
\usage{
dPlot(x, title = "", xlab = "score", ylab = "Frequency", ymax = 0, xmin = NULL)
}
\arguments{
\item{x}{data.frame object where the density of columnar values will be plotted}

\item{title}{Character indicating plot title}

\item{xlab}{Character indicating the x-axis label. Default is 'score'}

\item{ylab}{Character indicating the y-axis label. Default is 'Frequency'}

\item{ymax}{Numeric value indicating the maximum y-value. If values are potentially greater than ymax, ymax will be adjusted to relfect the new maximum (rounded to 6 decimal points).}

\item{xmin}{Numeric value indicating the minimum x-value. Default is NULL, meaning the x-axis limits will range from the negative maximum absolute x-value to the positive maximum absolue x-value.}
}
\value{
A plot with filled densities for each column.
}
\description{
Create a plot showing the densities of values in columns of a data frame
}
