% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSEA.R
\name{enPlot}
\alias{enPlot}
\title{Generate enrichment plot from GSEA results}
\usage{
enPlot(gseaRes, rnk, gmt, title = NULL)
}
\arguments{
\item{gseaRes}{Data frame output of GSEA function containing specific row(s) to make enrichment plots}

\item{rnk}{named vector of gene rankings (names are genes with matching nomenclature of geneset file). Must be same as one used to create gseaRes input}

\item{gmt}{named list of gene sets or character indicating path to gmt file used to generate gseaRes input}

\item{title}{character vector of length nrow(gseaRes) with a title for each plot. Leave NULL for automated titles (parsed by '\%' or ':")}
}
\value{
Enrichment plot with rank of genes on x-axis and running enrichment score on y-axis
}
\description{
Generate an enrichment plot for specific GSEA results
}
