% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Correlations.R
\name{expCor}
\alias{expCor}
\title{Expresion correlation plots of all gene expressions}
\usage{
expCor(
  counts,
  cond,
  title,
  method = "spearman",
  transform = "none",
  printInd = FALSE
)
}
\arguments{
\item{counts}{data.frame object with rows=genes and columns=samples of gene expression}

\item{cond}{character vector of length ncol(counts) assigning each sample to a condition}

\item{title}{character indicating the title of the plot(s)}

\item{method}{character indicating correlation method. Default is 'spearman'}

\item{transform}{character indicating if gene expression should be transformed. Default is "none", other options are "log2" and "log10"}

\item{printInd}{Boolean indicating if individual correlation plots should be printed in addition to paired plots. Default is FALSE.}
}
\value{
correlation plots
}
\description{
Expresion correlation plots of all gene expressions
}
