% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploreData.R
\name{exploreData}
\alias{exploreData}
\title{Explore your Data}
\usage{
exploreData(res, counts, cond)
}
\arguments{
\item{res}{A DESeq2 results object obtained from 'results(dds)' or a data.frame with the same column name values as a DESeq2 results object and rownames as genes. Can also be a named list of results objects.}

\item{counts}{Normalized count matrix with rows as genes and columns as samples. Can also be a named list of count objects - ensure same order and names as the results object list.}

\item{cond}{Character vector indicating conditions belonging to each sample (same order as colnames(counts)). Can also be a named list of condition vectors - ensure same order and names as results and counts object lists.}
}
\value{
A shiny app will be generated allowing you to explore your data interactively using BinfTools. Options to be selected:
Dataset: the name of the dataset to look at (from your named list of objects). If no list is provided, defaults to 'data'.
Control condition: The control condition for the current dataset (selected from cond)
Absolute log2FoldChange Threshold: The log2 fold-change threshold for differentially expressed genes (defaults to log2(1.5))
P-value column: Use adjusted p-value (padj) or raw p-value (pvalue)
P-value threshold: The significance threshold for differentially expressed genes (default=0.05)
Gene to colour/label: Type in gene name to colour/label on volcano plot, MA plot, and heatmap (must match gene symbols in res/counts objects)
Scale to point sizes: If yes, points in volcano plot and MA plot will be scaled to gene expression levels or significance, respectively
species: Indicate the species of the data for pathway enrichment analyses
Pathway Enrichment Sources: Select data source for pathway enrichment analysis
Print top 10 terms: sig= top 10 significant terms from pathway enrichment. enr= top 10 enriched terms from pathway enrichment
Pathway Keyword (GO only): Type in a key word for targeted analysis. Must have no spaces. Term names containing this string (even as part of a word) will be included for targeted analysis
Count plot scaling: "none" = use valuse as is, "log10"= plot as log10(1+normcounts), "zscore"= plot z-score normcounts
Count plot method: ind= use values from individual samples, mean= use average from samples in each condition, geoMean= use geometric mean from samples in each condition, median= use median from samples in each condition, perMean= percent mean of samples in each condition
}
\description{
This function generates a shiny App that allows you to interactively explore your data.
}
