% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Convert_res.R
\name{fromLimma}
\alias{fromLimma}
\title{Convert outputs from tools other than DESeq2 for compatibility with BinfTools}
\usage{
fromLimma(topTable, geneNames = rownames(topTable))
}
\arguments{
\item{topTable}{A data frame originating from the output of limma::topTable()}

\item{geneNames}{Acharacter vector specifying the names of the genes. Defaults to rownames(topTable) or rownames(topTags$table).}
}
\value{
A data frame of results compatible with BinfTools commands
}
\description{
These functions convert results from Limma or EdgeR for compaitibility with BinfTools
}
