% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BarGene.R
\name{geoMean}
\alias{geoMean}
\title{Calculate the geometric mean of a vector of numeric values}
\usage{
geoMean(a)
}
\arguments{
\item{a}{numeric vector}
}
\value{
geometric mean
}
\description{
Calculate the geometric mean of a vector of numeric values
}
