% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSym.R
\name{getSym}
\alias{getSym}
\title{Convert to gene symbols}
\usage{
getSym(
  object,
  obType = c("res", "counts"),
  species = "hsapiens",
  target = "ENSG",
  addCol = F
)
}
\arguments{
\item{object}{a data.frame object (results or normalized counts) with the rownames as the gene IDs to convert}

\item{obType}{Character vector (one of 'res' or 'counts') indicating whether \emph{object} is a results object, or normalized counts object, respectively.}

\item{species}{Character vector indicating the species (e.g. "hsapiens" (\emph{default}), "mmusculus").}

\item{target}{Character vector indicating the current format of the gene IDs (e.g. "ENSG", "RefSeq","Entrez", see https://biit.cs.ut.ee/gprofiler/convert for all options). If IDs are ENSEMBL gene IDs with version nmbers (i.e. "ENSG0000****.1"), used "ENSGV".}

\item{addCol}{Boolean indicating if an additional column named "SYMBOL" should be added, leaving the rownames as the original format. If FALSE (\emph{default}), rownames of \emph{object} will be replaced with gene symbols - duplicate gene symbols will be handled by keeping the symbol with the highest gene expression values.}
}
\value{
A data.frame object in the format indicated by \emph{obType} with gene symbols as the rownames (if \emph{addCol}=F) or gene symbols added as their own column named "SYMBOL".
}
\description{
This function makes use of gprofiler2's 'gconvert' function to convert gene
ids in the rownames of the results object or the normalized counts object to
a gene symbol for easy labeling in plots.
}
\details{
This will be integrated into the volcanoPlot(), MA_Plot(), and zheat() functions
in the future.
}
