% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmt.R
\name{gmtHeat}
\alias{gmtHeat}
\title{Make a heatmap of gene expression grouped by gene sets}
\usage{
gmtHeat(
  counts,
  cond,
  gmt,
  con = NULL,
  labgenes = NULL,
  title = "",
  avgExp = TRUE,
  zscore = TRUE,
  colAnno = NULL,
  hmcol = NULL,
  intClus = TRUE,
  printEach = FALSE,
  retGroups = FALSE
)
}
\arguments{
\item{counts}{data frame of normalized gene expression values}

\item{cond}{character vector indicating which condition each column of counts belongs to}

\item{gmt}{named list of gene sets to be used in the heatmap. Genes must be found in rownames(counts).}

\item{con}{character indicating the control condition from cond. Or the order in which conditions in cond should appear on the heatmap.}

\item{labgenes}{character indicating which genes (if any) should be labeled on the heatmap. Default NULL will label all genes. Set to "" to label no genes.}

\item{title}{character indicating the title of the resulting heatmap}

\item{avgExp}{Boolean indicating if gene expression should be averaged within each condition (TRUE) or if each individual replicate should be plotted (FALSE; default).}

\item{zscore}{Boolean indicating if gene expression should be z-score scaled (TRUE; default) or not (FALSE).}

\item{colAnno}{Dataframe with rownames=colnames counts and at least one column to annotate the columns of the heatmap. Leave NULL for no annotation.}

\item{hmcol}{colorRampPalette object of length 100 indicating colour scheme of heatmap. Leave NULL for default colours.}

\item{intClus}{Boolean indicating if hierarchical clustering should be performed within each gene set. Note trees will not be shown. Default = TRUE. If FALSE, heatmap will be presented in the order in which genes appear in the gene sets.}

\item{printEach}{Boolean indicating if heatmaps for each individual gene set should be printed. Only works when intClus=TRUE. Trees will be shown. Default = FALSE.}

\item{retGroups}{Boolean indicating if named list of data frames of gene expression subset to each gene set should be returned (z-score normalized if zscore=TRUE). Default is FALSE. if TRUE, heatmap won't be plotted.}
}
\value{
Annotated heatmap of gene expression of all gene sets provided or named list of data frames of gene expression subset to each gene set.
}
\description{
Make a heatmap of gene expression grouped by gene sets
}
