% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GO.R
\name{GO_Dot}
\alias{GO_Dot}
\title{Create a dot plot of GO results}
\usage{
GO_Dot(GOres, title = "")
}
\arguments{
\item{GOres}{Single results table from GO_GEM() output with returnRes=T}

\item{title}{Character indicating the title of the plot}
}
\value{
Dot plot representing enrichment, precision (as 'Gene Ratio'), and significance of enriched GO terms.
}
\description{
Create a dot plot of GO results
}
