% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GO.R
\name{GO_plot}
\alias{GO_plot}
\title{Generates GO figures}
\usage{
GO_plot(GOres, prefix, ts, pdf, fig, print = c("both", "sig", "enr"))
}
\description{
Function that generates the top 10 term figures for GO_GEM()
}
