% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GO.R
\name{GOgrep}
\alias{GOgrep}
\title{Grep a key word from a GO results table and return specific columns}
\usage{
GOgrep(GOtable, key, cols = c("term_name", "intersection"))
}
\arguments{
\item{GOtable}{Data frame returned from GO_GEM or clusGO with returnRes=T}

\item{key}{Character to be matched in the column 'term_name'. Note that the key is not case sensitive and will return parital matches. Be sure to review the results.}

\item{cols}{Character vector of any of colnames(GOtable) indicating which columns should be returned. Default is c("term_name","intersection").}
}
\value{
Data frame of subset GOtable where key is found in term_name and columns match cols. If no matches are found, NA is returned.
}
\description{
Grep a key word from a GO results table and return specific columns
}
