% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GO.R
\name{GOHeat2}
\alias{GOHeat2}
\title{Dot heatmap for GO results}
\usage{
GOHeat2(GOresList, termList, ret = FALSE)
}
\arguments{
\item{GOresList}{Named list containing GO_GEM() results tables when returnRes=T}

\item{termList}{Named list of relevant GO term names - made with makeTermList()}

\item{ret}{Boolean indicating if values used for heatmap should be returned. Default is FALSE.}
}
\value{
Heatmap figure representing enrichment and precision (as 'Gene Ratio') of selected enriched pathways.
}
\description{
Dot heatmap for GO results
}
