% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSEA.R
\name{GSEA}
\alias{GSEA}
\title{Gene Set Enrichment Analysis}
\usage{
GSEA(
  rnk,
  gmt,
  pval = 1,
  ts = c(10, 600),
  nperm = 10000,
  parseBader = TRUE,
  plot.title = "Gene Set Enrichment Analysis",
  useGAGE = TRUE
)
}
\arguments{
\item{rnk}{named vector of gene rankings (names are genes with matching nomenclature of geneset file)}

\item{gmt}{named list of gene sets or character indicating gmt file path for gene sets to be used in GSEA analysis}

\item{pval}{Adjusted p-value threshold to filter results. Set to 1 to return all results. Default=1}

\item{ts}{numeric vector of length two indicating the minimum and maximum gene set sizes}

\item{nperm}{integer indicating number of permutations to run. Default is 10000}

\item{parseBader}{Boolean indicating if gene set names should be parsed for the output figure following Bader Lab nomenclature (i.e. '\%' as delimiter)}

\item{plot.title}{Title for the plot}

\item{useGAGE}{Boolean indicating if GAGE results should be used to filter fgsea results. Default=TRUE}
}
\value{
A data frame of GSEA results and a figure showing normalized enrichment scores (NES) for top positive and negative enriched gene sets.
}
\description{
Perform a GSEA using fgsea and gage analyses
}
