% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSEA.R
\name{GSEA_EM}
\alias{GSEA_EM}
\title{Make Enrichment map files}
\usage{
GSEA_EM(gsea, rnk, prefix = NULL, gmt = NULL)
}
\arguments{
\item{gsea}{To be documented!}

\item{rnk}{To be documented!}

\item{prefix}{To be documented!}

\item{gmt}{To be documented!}
}
\value{
To be documented!
}
\description{
Make Enrichment map files
}
