% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSEA.R
\name{gsea_gmt}
\alias{gsea_gmt}
\title{Make a custom gmt from GSEA terms}
\usage{
gsea_gmt(terms, gmt, leadingEdge = FALSE)
}
\arguments{
\item{terms}{Character vector of GSEA terms (pathway column from results of GSEA() function)}

\item{gmt}{GMT file name or R object used to generate the gsea results using GSEA(). Can be GSEA results object if setting leadingEdge=TRUE.}

\item{leadingEdge}{Boolean indicating if the leading edge genes only should be extracted. Default=FALSE}
}
\value{
List gmt object of genesets with terms from GSEA analysis
}
\description{
Make a custom gmt from GSEA terms
}
