% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenerateGSEA.R
\name{GSEAclus}
\alias{GSEAclus}
\title{Create clusters of terms for Enrichment Map}
\usage{
GSEAclus(gsFile, term, clusName, dir, retRNK = FALSE)
}
\arguments{
\item{gsFile}{gmt file contatining gene sets used for GSEA}

\item{term}{character vector of term keys to use to identify terms belonging to each cluster}

\item{clusName}{character vector of same length as 'term' in the same order telling which cluster each term belongs to}

\item{dir}{output directory for tsv file. Will be named "GSEA_clusters.tsv"}

\item{retRNK}{Boolean. If TRUE, return the data frame containing the terms and clusters. Default is FALSE.}
}
\description{
Create clusters of terms for Enrichment Map
}
