% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zheat.R
\name{heatClusRes}
\alias{heatClusRes}
\title{Add clusters to results or count matrix}
\usage{
heatClusRes(resList, heatClus)
}
\arguments{
\item{resList}{A list or single data frame of DESeq2 results or gene expression counts}

\item{heatClus}{The data frame with rownames=genes and a column named Cluster. The output of heatClus}
}
\value{
The original object(s) with a column named 'cluster' - for use with clusRelev()
}
\description{
Make a data frame of DESeq2 results or gene expression counts that shows the cluster each gene belongs to
}
