% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GO_Tables.R
\name{listGO_Table}
\alias{listGO_Table}
\title{GO Table from list of GO results}
\usage{
listGO_Table(
  GOres,
  title = NULL,
  subtitle = NULL,
  sig = T,
  ts = c(10, 500),
  colnames = c("Term", "Enrichment", "FDR"),
  numTerm = 10,
  retGT = F,
  printGT = T
)
}
\arguments{
\item{GOres}{List of GO results table returned from GO_GEM or clusGO when returnRes=T.}

\item{title}{Character for Table title, leave NULL to avoid a title}

\item{subtitle}{Character for subtitle of table. Leave NULL to not have subtitle}

\item{sig}{Boolean indeicating if terms should be ordered by significance. FALSE lets terms be ordered by enrichment. Default=T.}

\item{ts}{numeric vector of length two indicating minimum and maximum terms size cutoffs for enrichment and significance, respectively. Default=c(10,500)}

\item{colnames}{Character vector of length 3 indicating the output column names for the Term name, enrichment, and adjusted p-value. Defaults to c("Term", "Enrichment", "FDR")}

\item{numTerm}{Number indicating the number of terms to print in the table. Default=10.}

\item{retGT}{Boolean indicating if the gt table object should be returned - useful for custom formatting. Default=F.}

\item{printGT}{Boolean indicating if the gt table object should be printed to Viewer. Default =T.}
}
\value{
A table of class gt of the top GO results for each entry in the list of GO results.
}
\description{
Make a publication-quality table from your GO results
}
