% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadData.R
\name{loadData}
\alias{loadData}
\title{Load HTSeq-count data from a sampleTable into a countMatrix and colData object}
\usage{
loadData(sampleTable)
}
\arguments{
\item{sampleTable}{Character indicating the path to a tab-delimited (.txt) or comma-separated (.csv) sampleTable, or a data frame.}
}
\value{
A list of length 2. Counts = count matrix and colData = metadata for samples
}
\description{
Function to use a sampleTable (same format for DESeqDataSetFromHTSeqCount) to generate a count matrix and colData object
}
