% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadData.R
\name{loadFC}
\alias{loadFC}
\title{Function to load featureCounts data}
\usage{
loadFC(filename)
}
\arguments{
\item{filename}{Character indicating the path to the featureCounts table to import.}
}
\value{
 A list of length 2: Counts=featureCount count matrix and annotation= Gene Annotation information from featureCounts
\description{
Function to load featureCounts data
}
