% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MA_plot.R
\name{MA_Plot}
\alias{MA_Plot}
\title{Custom MA Plot}
\usage{
MA_Plot(
  res,
  title,
  p = NULL,
  pval = NULL,
  FC = 1,
  lab = NULL,
  col = NULL,
  fclim = NULL,
  showNum = TRUE,
  returnDEG = F,
  sigScale = F
)
}
\arguments{
\item{res}{A DESeq2 results object obtained from 'results(dds)' or a data.frame with the same column name values as a DESeq2 results object and rownames as genes}

\item{title}{A character vector indicating the title of the plot}

\item{p}{A number indicating the threshold for 'padj' where padj<p are significant genes. Should not be used if using 'pval'}

\item{pval}{A number indicating the threshold for 'pvalue' where pvalue<pval are significant genes. Should not be used if using 'p'}

\item{FC}{A number indicating the log2FoldChange threshold where abs(log2FC)>FC are significant genes. Default is 1 - can be 0 if not using fold-change threshold.}

\item{lab}{A character vector of genes to be labeled on the plot. Should correspond with rownames(res). Default is NULL. Enter 'labDEG' to label all DEGs given thresholds.}

\item{col}{A character vector of genes to be coloured orange on the plot. Should correspond with rownames(res). Default is NULL.}

\item{fclim}{A number indicating the maximum log2FoldChange value desired on the volcano plot (x-axis limits). Points outside this limit will appear as diamonds on the limits.}

\item{showNum}{A boolean indicating whether gene numbers should be displayed on the plot. Default is TRUE.}

\item{returnDEG}{A boolean indicating whether DEGs (using given thresholds) should be returned as a list (down, up)}

\item{sigScale}{A boolean indicating if the point size should be scaled by significance (more significant = larger point). Default=FALSE.}
}
\value{
An MA plot with x-axis indicating gene expression 'baseMean' and y-axis indicating log2FoldChange. Blue dots are downregulated genes, red dots are upregulated genes.
}
\description{
This function creates an MA plot from a DESeq2 results object. It is like
'plotMA()' from DESeq2, but more customizable.
}
