% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{mgrep}
\alias{mgrep}
\title{grep for multiple patterns}
\usage{
mgrep(p, x, value = F, ic = F, v = F, unique = F)
}
\arguments{
\item{p}{Character vector of regex patterns to search}

\item{x}{Target vector to search for patterns}

\item{value}{Boolean indicating if the value of matching patterns should be returned. If FALSE, only the index of matching patterns will be returned. Default is FALSE.}

\item{ic}{Boolean indicating if case should be ignored. Default is FALSE.}

\item{v}{Boolean indicating if the returned values should be inverted. Default is FALSE.}

\item{unique}{Boolean indicating if duplicate values should be removed before returning. Default is FALSE.}
}
\value{
Vector of indices of x that contain matching patterns of p. If value = TRUE, values of x containing p will be returned.
}
\description{
Search for multiple regex patterns in a given vector
}
