% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BarGene.R
\name{normToGene}
\alias{normToGene}
\title{Normalize gene expression to specific genes}
\usage{
normToGene(counts, norm)
}
\arguments{
\item{counts}{count matrix (normalized or raw) with genes as rows and samples as columns}

\item{norm}{character vector of length >=1 of gene names (matching rownames(counts)) to use as reference genes for normalization. If length > 1, geometric mean of genes' expressions will be used.}
}
\value{
data frame of dim(counts) of gene expression normalized relative to 'norm'
}
\description{
This function will normalize gene expression in a count matrix relative to a given gene
or vector of genes.
}
