% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PathWC.R
\name{PathWC}
\alias{PathWC}
\title{Make a Word Cloud of enriched pathway terms}
\usage{
PathWC(
  text,
  cols = "Dark2",
  retTerms = F,
  minfreq = 3,
  rmwords = c("regulation", "process", "positive", "negative", "mediated", "cell",
    "cellular", "protein")
)
}
\arguments{
\item{text}{Character vector of terms to be processed for a word cloud}

\item{cols}{Colour palette. Can be RColourBrewer palette name, rgb, colour names, or hexadecimal. Default is "Dark2"}

\item{retTerms}{Boolean if term-document matrix dataframe should be returned. Default=F}

\item{minfreq}{Minimum frequency for a word to be included in the word cloud. Default=3}

\item{rmwords}{Character vector of words to be filtered out (in addition to the default english stop words and words shorter than 3 letters) - defaults to c("regulation","process","positive","negative","mediated", "cell","cellular", "protein")}
}
\value{
A Word cloud showing most frequent words in input text, and if indicated, a data frame of word frequencies.
}
\description{
Use the textmining and wordcloud packages to make a word cloud. Great for using
with enriched GO/GSEA term names
}
