% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusFigs.R
\name{plotClusBox}
\alias{plotClusBox}
\title{Make a boxplot grouped by gene sets or clusters}
\usage{
plotClusBox(x, yax = "Log2FoldChange", col = "Dark2", title = "", showStat = T)
}
\arguments{
\item{x}{Data frame or named list of data frames of average gene expression or log2FoldChanges of a subset of genes, where each column is a condition or contrast (in the order you want displayed), and rows are genes.}

\item{yax}{Character indicating the y-axis label of the boxplot. Default="Log2FoldChange".}

\item{col}{Character indicating the RColorBrewer palette name or list of colours (hex, name, rgb()) to be used for the bar plot. Default is "Dark2"}

\item{title}{Character indicating the title of the plot}

\item{showStat}{Boolean indicating if significance of pairwise contrasts within each group should be shown. Default=T}
}
\value{
Boxplot or grouped boxplot and tibble contatining statistics from pairwise t-tests.
}
\description{
Make a boxplot grouped by gene sets or clusters
}
