% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Correlations.R
\name{plotCor}
\alias{plotCor}
\title{Correlation Plotting
Plot a correlation bewteen any two columns from any data frame(s) or named vectors}
\usage{
plotCor(x, y, xCol = F, yCol = F, xlab = "", ylab = "", title = "", scale = F)
}
\arguments{
\item{x}{Data frame with rownames or named vector containing data to be plotted on the x-axis}

\item{y}{Data frame with rownames or named vector containing data to be plotted on the y-axis}

\item{xCol}{Character indicating column name in x of data to be plotted on the x-axis. Leave FALSE if x is a named vector.}

\item{yCol}{Character indicating column name in y of data to be plotted on the y-axis. Leave FALSE if y is a named vector.}

\item{xlab}{Character indicating x-axis label}

\item{ylab}{Character indicating y-axis label}

\item{title}{Character indicating title of plot}

\item{scale}{Should the data be z-score scaled? Default is FALSE.}
}
\value{
Scatter plot of data with line of best fit and R-value of correlation. Note that x and y data are paired based on the rownames of the data frame or names of the named vector. Non-paired values will not be plotted.
}
\description{
Correlation Plotting
Plot a correlation bewteen any two columns from any data frame(s) or named vectors
}
