% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenerateGSEA.R
\name{plotrnk}
\alias{plotrnk}
\title{Plot the distribution of scores by gene rank}
\usage{
plotrnk(rnk)
}
\arguments{
\item{rnk}{rnk object obtained from GeneratedGSEA() with retRNK=T}
}
\value{
Plot of scores in ranked gene list
}
\description{
Plot the distribution of scores by gene rank
}
