% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTime.R
\name{plotTime}
\alias{plotTime}
\title{Plot time course data}
\usage{
plotTime(
  genes,
  counts,
  condition,
  timePoints,
  tOrd,
  title = "",
  cols = "Dark2",
  yax = "Expression",
  xax = "Time Point",
  eb = "se",
  showStat = T
)
}
\arguments{
\item{genes}{character vector of at least one gene found in rownames(counts)}

\item{counts}{data frame of gene expression where rows are genes and columns are samples}

\item{condition}{character vector describing the condition of each sample in the same order of colnames(counts)}

\item{timePoints}{character vector describing the time points of each sample in the same order of colnames(counts)}

\item{tOrd}{character vector describing the order in which the timePoints should be plotted. Must be length of length(unique(timePoints))}

\item{title}{character describing the title of the plot}

\item{cols}{Character indicating the RColorBrewer palette name or list of colours (hex, name, rgb()) to be used. Default is "Dark2"}

\item{yax}{character indicating how the y-axis should be labeled - note that error bar notation will be added automatically if applicable}

\item{xax}{character indicating x-axis label}

\item{showStat}{Boolean indicating if significance should be shown on the plot.}
}
\value{
plot showing expression changes over time
}
\description{
This function accepts a single gene or vector of multiple genes along with counts, conditions, and timepoints
to plot the difference of gene expression between conditions over time.
}
