% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotVenn.R
\name{plotVenn}
\alias{plotVenn}
\title{Make a Venn diagram}
\usage{
plotVenn(
  x,
  title = "Venn Diagram",
  cols = "Dark2",
  lty = "blank",
  scale = FALSE,
  retVals = FALSE
)
}
\arguments{
\item{x}{A named list (max length of 5) of characters (genes) to plot in a Venn Diagram}

\item{title}{Character indicating the title of the plot. Default="Venn Diagram"}

\item{cols}{Character indicating the RColorBrewer palette name or list of colours (hex, name, rgb()) to be used. Default is "Dark2"}

\item{lty}{Line type (1=solid line, 2=dashed line, default="blank")}

\item{scale}{Boolean indicating if circles should be scaled to size (works only for 2 or 3-way Venn diagrams). Default=FALSE.}

\item{retVals}{Boolean indicating if the overlaps should be returned (default=FALSE)}
}
\value{
An image of a Venn diagram showing overlaps between groups
}
\description{
Make a Venn diagram from a named list of genes. Uses functions from the VennDiagram package.
}
