% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MA_plot.R
\name{pseudoMA}
\alias{pseudoMA}
\title{Make a 'pseudo' MA plot
Make a MA plot even when you have only n=1. Note that this can be used if one or both condition groups have more than one or more replicates.}
\usage{
pseudoMA(counts, treat, con, FoldChange = 1, title = "", retDA = F, retRes = F)
}
\arguments{
\item{counts}{data frame of normalized counts to be used.}

\item{treat}{character vector of column name(s) in counts to be used as the treatment condition.}

\item{con}{character vector of column name(s) in counts to be used as the control/reference condition.}

\item{FoldChange}{absolute log2FoldChange threshold (treat/con) to subset the 'differently activated' genes - Note no statistics are performed with this, and so these genes cannot be named as 'differential'}

\item{title}{Character indicating the title of the plot}

\item{retDA}{Boolean indicating if differently activated genes should be returned as a list (Down, Up). Default is FALSE.}

\item{retRes}{Boolean indicating if 'pseudo' results object should be returned. Default is FALSE, and will not return if retDA=TRUE. Note that any p-values indicated in this table are dummy values.}
}
\value{
MA plot with differently activated genes highlighted, named list of differently activated genes if retDA=T, and 'pseudo' results object if retRes=T (and retDA=F).
}
\description{
Make a 'pseudo' MA plot
Make a MA plot even when you have only n=1. Note that this can be used if one or both condition groups have more than one or more replicates.
}
