% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadData.R
\name{quart_group}
\alias{quart_group}
\title{Generate and assign quantiles
Function great for making a 'condition' vector based on gene expression values}
\usage{
quart_group(x, groups = c("1st", "2nd", "3rd", "4th"))
}
\arguments{
\item{x}{numeric vector to be used to generate quantiles}

\item{groups}{character vector of names to be assigned to quantiles in order of ascending values. Number of quantiles generated is based off of length(groups).}
}
\value{
character vector of length(x) with names of groups assigned to each value of x
}
\description{
Generate and assign quantiles
Function great for making a 'condition' vector based on gene expression values
}
