% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmt.R
\name{readGMT}
\alias{readGMT}
\title{Read a gmt file into R}
\usage{
readGMT(gmt)
}
\arguments{
\item{gmt}{Character with path to file in .gmt format}
}
\description{
This function will read a gmt file into a list object in R
}
