% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colPal.R
\name{showPals}
\alias{showPals}
\title{Show colour palettes from RColorBrewer}
\usage{
showPals()
}
\value{
Image of colour palettes from RColorBrewer with names
}
\description{
Show colour palettes from RColorBrewer
}
