% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSEA.R
\name{subGSEA}
\alias{subGSEA}
\title{Subset GSEA results}
\usage{
subGSEA(x, p = 0.05, n = 100, byNES = TRUE)
}
\arguments{
\item{x}{GSEA results produced by GSEA() function}

\item{p}{FDR threshold to filter GSEA results. Default = 0.05.}

\item{n}{Number of top terms to return. Default is 100. Set to Inf to return all possible filtered results.}

\item{byNES}{Boolean indicating how terms should be ordered. If TRUE (default), terms are ordered by decreasing absolute normalized enrichment score. If FALSE, terms are ordered by decreasing significance.}
}
\description{
Subset GSEA results
}
