% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusFigs.R
\name{subRes}
\alias{subRes}
\title{Subset results objects using various parameters}
\usage{
subRes(res, genes = NULL, p = NULL, pval = NULL, FC = NULL)
}
\arguments{
\item{genes}{character vector of genes in rownames(res) to subset. Default=NULL.}

\item{p}{adjusted p-value threshold. Genes must have padj<p to be returned. Default=NULL.}

\item{pval}{raw p-value threshold. Genes must have pvalue<pval to be returned. Default=NULL.}

\item{FC}{absolute log2FoldChange threshold. Genes must have |log2FoldChange|>FC to be returned. Default=NULL.}

\item{results}{data frame}
}
\value{
Data frame subset to genes matching any and all of the provided parametes.
}
\description{
Subset results objects using various parameters
}
