% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcanoPlot.R
\name{vol3d}
\alias{vol3d}
\title{Make a 3D volcano plot
combine 2 results objects into one volcano plot - note this is experimental and not usually done}
\usage{
vol3d(
  x,
  y,
  xlab = "res1",
  ylab = "res2",
  pval = 0.05,
  pcol = "pvalue",
  usep = NULL
)
}
\arguments{
\item{x}{DESeq2 results object to be plotted for the x-axis}

\item{y}{DESeq2 results object to be plotted for the y-axis}

\item{xlab}{Contrast name for x. Default='res1'.}

\item{ylab}{Contrast name for y. Default='res2'.}

\item{pval}{significance threshold for differential genes. Default is 0.05}

\item{pcol}{Character of either 'pvalue' or 'padj' indicating which column to use to plot significance. Default is "pvalue".}

\item{usep}{numeric of either 1 or 2 indicating which p-values to plot (x or y, respectively). If left NULL (default). P-values from both x and y are combined using Fisher's combined probability test, and the resulting p-values are plotted.}
}
\value{
3D volcano plot with every combination of up/down-regulated genes labeled and a data frame of the combined data used to generate the plot.
}
\description{
Make a 3D volcano plot
combine 2 results objects into one volcano plot - note this is experimental and not usually done
}
