% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmt.R
\name{writeGMT}
\alias{writeGMT}
\title{Write a gene set to file in gmt format}
\usage{
writeGMT(geneSet, filename)
}
\arguments{
\item{geneSet}{The geneset list object (generated by customGMT())}

\item{filename}{The output filename for the gmt file. Use extension .gmt}
}
\description{
This function will accept a list of genes belonging to a gene set and write
it to a given filename in gmt format.
}
