% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zheat.R
\name{zheat}
\alias{zheat}
\title{Z-score normalized heatmap of gene expression}
\usage{
zheat(
  genes = NULL,
  counts,
  conditions,
  con = "WT",
  title = "DEGs",
  labgenes = NULL,
  zscore = T
)
}
\arguments{
\item{genes}{A character vector of genes to subset from rownames(counts) to create the heatmap}

\item{counts}{Normalized count matrix with rows as genes and columns as samples. Use 'counts(dds, normalized=T)'.}

\item{conditions}{Character vector indicating conditions belonging to each sample (same order as colnames(counts))}

\item{con}{Character vector indicating the control condition level in 'conditions'. Default is "WT".}

\item{title}{Character vector indicating the title of the figure}

\item{labgenes}{Character vector corresponding to rownames(counts) of genes to be labelled on the side of the heatmap. Leave as NULL to label all genes. Use "" to label no genes.}

\item{zscore}{Boolean indicating if counts should be z-score normalized (default=T)}
}
\value{
Heatmap of z-score normalized gene expression with control condition samples appearing first.
}
\description{
A function to generate a z-score normalized heatmap of gene expression from
a DESeq2 counts object.
}
