% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zheat.R
\name{zheat_v2}
\alias{zheat_v2}
\title{Z-score normalized heatmap of gene expression}
\usage{
zheat_v2(
  genes = NULL,
  counts,
  conditions,
  order.by = NULL,
  sort.genes.by = NULL,
  title = "DEGs",
  labgenes = NULL,
  zscore = T,
  avgExp = F,
  rclus = F,
  hmcol = NULL,
  retClus = F,
  annoCols = "Dark2",
  border_color = "grey60"
)
}
\arguments{
\item{genes}{A character vector of genes to subset from rownames(counts) to create the heatmap}

\item{counts}{Normalized count matrix with rows as genes and columns as samples. Use 'counts(dds, normalized=T)'.}

\item{conditions}{Character vector indicating conditions belonging to each sample (same order as colnames(counts))}

\item{order.by}{Character vector indicating the arbitrary appearance order of the conditions level. Default is the default order of levels(conditions).}

\item{sort.genes.by}{Character vector indicating the samples for sorting heatmap rows (genes).}

\item{title}{Character vector indicating the title of the figure}

\item{labgenes}{Character vector corresponding to rownames(counts) of genes to be labelled on the side of the heatmap. Leave as NULL to label all genes. Use "" to label no genes.}

\item{zscore}{Boolean indicating if counts should be z-score normalized (default=T)}

\item{avgExp}{Boolean indicating if values should be averaged within each condition (i.e. plot one value per condition). Default=F.}

\item{rclus}{Boolean indicating if rows (genes) should be clustered. Leave FALSE for genes to be ordered on decreasing expression in 'con'.}

\item{hmcol}{Color ramp palette of length 100 for custom heatmap colouring. Leave NULL for defaults.}

\item{retClus}{Boolean indicating if clustered row names should be returnd. Default=FALSE.}

\item{annoCols}{Character vector indicating RColorBrewer palette name or colours for annotation colours if rclus is provided a data.frame.}

\item{border_color}{color of cell borders on heatmap, use NA if no border should be drawn.}
}
\value{
Heatmap of z-score normalized gene expression with control condition samples appearing first.
}
\description{
A function to generate a z-score normalized heatmap of gene expression from
a DESeq2 counts object.
}
