% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BarGene.R
\name{barGene}
\alias{barGene}
\title{Create a bar plot of gene expression}
\usage{
barGene(
  genes,
  counts,
  conditions,
  title = "Gene expression",
  norm = NULL,
  eb = "sd",
  returnDat = F,
  col = "Dark2",
  ord = NULL,
  con = NULL,
  stat.test = "t.test",
  hide.ns = T
)
}
\arguments{
\item{genes}{character vector of genes to plot. Must match rownames(counts)}

\item{counts}{count matrix of gene expression, where rows are genes and columns are samples}

\item{conditions}{character vector of length ncol(counts) describing the condition of each sampmle in counts}

\item{title}{character describing the title of the plot}

\item{norm}{character describing the condition to use as a reference for relative gene expression. Each gene's expression will be set relative to this condition. Leave 'NULL' if plotting raw values.}

\item{eb}{character describing the style of error bar. "sd" = standard deviation, "se" = standard error of the mean. Use '0' if no error bars to be plotted. Default is 'sd'.}

\item{returnDat}{Boolean indicating if list of raw and summarized data should be returned for further analysis. Default is FALSE.}

\item{col}{character indicating the RColorBrewer palette name or list of colours (hex, name, rgb()) to be used. Default is "Dark2"}

\item{ord}{character indicating the order in which the samples should appear (overrides any ordering from using 'norm' argument). Default is NULL.}

\item{con}{character indicating control condition if pairswise t-tests are to be performed. Leave NULL to not include stats. Set to 'show.all' to show all pairwise comparisons within each gene.}

\item{stat.test}{character indicating either "t.test" or "wilcox" for stats when 'con' is defined. Default is "t.test"}

\item{hide.ns}{logical value. If TRUE, hide ns symbol when displaying significance levels.}
}
\value{
Bar plot of gene expression and list of length 2 containing 'rawData' and 'Summary' of gene expression data if 'returnDat' is TRUE.
}
\description{
This function will accept a list of genes and counts to produce a bar plot of gene expression in different conditions
}
