% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusGO.R
\name{clusGO}
\alias{clusGO}
\title{Pathway Enrichment Analysis on Clusters}
\usage{
clusGO(
  clusRes,
  species = "hsapiens",
  bg = rownames(clusRes),
  source = NULL,
  corr = "fdr",
  iea = FALSE,
  prefix = "ClusGO",
  ts = 10,
  pdf = T,
  fig = T,
  returnGost = F
)
}
\arguments{
\item{clusRes}{data.frame object that is output from clusFigs()}

\item{...}{Arguments for GO_GEM() see ?GO_GEM for details. \emph{bg} is automatically set as rownames(clusRes).}
}
\value{
For each cluster, output from GO_GEM() is generated and output to "\emph{prefix}\emph{cluster}\emph{n}"
}
\description{
Run Pathway Enrichment Analysis (same as GO_GEM()) on output from clusFigs()
}
