% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GO.R
\name{filtGO}
\alias{filtGO}
\title{Filter a list of GO results to remove/replace tables with no enriched terms}
\usage{
filtGO(x, replace = F)
}
\arguments{
\item{x}{list of GO results tables returned when returnRes=T in GO_GEM() or clusGO()}

\item{replace}{Boolean indicating if slots with no enriched GO terms should be replaced with a 'dummy table' - this is good if you want to run a GOHeat analysis. Default is FALSE.}
}
\value{
list of GO results tables with tables containing no enriched terms being removed or replaced.
}
\description{
Filter a list of GO results to remove/replace tables with no enriched terms
}
