% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GO.R
\name{GO_GEM}
\alias{GO_GEM}
\title{Run g:profiler enrichment analysis}
\usage{
GO_GEM(
  geneList,
  species = "hsapiens",
  bg = NULL,
  source = NULL,
  corr = "fdr",
  iea = FALSE,
  prefix = "GO_analysis",
  ts = c(10, 500),
  pdf = T,
  fig = T,
  figCols = c("blue", "orange"),
  returnGost = F,
  writeRes = T,
  writeGem = T,
  returnRes = F
)
}
\arguments{
\item{geneList}{A character vector or named list of genes to query. If a named list is provided, the name of each list element is appended to the prefix for GO output.}

\item{species}{A character vector of the organism name to use. Concatenate the first letter of the name and family name. Default is 'hsapiens'}

\item{bg}{A character vector describing the genes making up the background. Default NULL}

\item{source}{A vector of data sources to use. Currently: GO (GO:BP, GO:MF, GO:CC), KEGG, REAC, TF, MIRNA, CORUM, HP, HPA, WP.}

\item{corr}{A character vector describing the correction method to use. One of 'gSCS', 'fdr', or 'bonferroni'. Default is 'fdr'}

\item{iea}{Boolean values indicating if electronic annotations should be excluded. Default FALSE.}

\item{prefix}{A character vector describing the path and prefix of the output files (should not include any file extensions)}

\item{ts}{Vector of length 2 indicating the minimum and maximum term size - the minimum/maximum number of genes per term when generating the plot of most enriched/significant terms. Default is c(10,500).}

\item{pdf}{Boolean indicating if bar plots should be exported to pdf. Default is TRUE.}

\item{fig}{Boolean indicating if bar plots should be printed to R output. Default is TRUE.}
  
\item{figCols}{Character indicating the RColorBrewer palette name or list of colours (hex, name, rgb()) to be used for enrichment and significance, respectively in the output figures. Default is c("blue","orange").}

\item{returnGost}{Boolean indicating if gost results should be returned for future use with gprofiler2 functions. Default is FALSE.}

\item{writeRes}{Boolean indicating if GO.txt results should be written to file 'prefix.GO.txt'}

\item{writeGem}{Boolean indicating if gem.txt results should be written to file.}
}
\value{
Exports a table of analysis results in 'prefix.GO.txt', a gem file in 'prefix.gem', and a pdf with two figures: top ten enriched terms followed by top ten significant terms in 'prefix.top10.pdf'
}
\description{
This function runs a g:profiler gost enrichment analysis on a query list of genes
and export a complete results table, a .gem file for use with EnrichmentMap Cytoscape app,
and a pdf of the top enriched and significant terms.
}
