% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssGSEA.R
\name{gsva_plot}
\alias{gsva_plot}
\title{Run ssGSEA and make a plot}
\usage{
gsva_plot(
  counts,
  geneset,
  method = "ssgsea",
  condition,
  con = NULL,
  title = "ssGSEA",
  compare = NULL,
  col = "Dark2",
  style = "violin"
)
}
\arguments{
\item{counts}{Normalized count matrix - preferably z-score normalized by row. E.g. t(scale(t(counts(dds, normalized=T))))}

\item{geneset}{List with length > 1 of character vectors indicating gene sets of interest. Can be created using qusage::read.gmt("geneset.gmt")}

\item{method}{Method to employ in estimation of gene-set enrichment scores. One of "ssgsea","gsva","zscore","plage". Default is "ssgsea}

\item{condition}{Character vector of conditions in the same order they appear as columns in 'counts'.}

\item{con}{Character indicating the control condition (Condition to be plotted first). Default is NULL.}

\item{title}{Character vector indicating the title of the plot}

\item{compare}{List of character vectors (each of length 2) indicating the pairwise comparisons to be made between conditions. Leave NULL for all comparisons to be made}

\item{col}{Character indicating the RColorBrewer palette name or list of colours (hex, name, rgb()) to be used. Default is "Dark2".}

\item{style}{Character indicating the style of plot ("violin" or "box"). Defaults to "violin".}
}
\value{
Violin or box plot of normalized enrichment scores for genesets between conditions
}
\description{
A function to run an ssGSEA analysis using normalized counts and create
a plot with pairwise comparisons
}
