% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssGSEA.R
\name{gsva_plot}
\alias{gsva_plot}
\title{Run ssGSEA and make a plot}
\usage{
gsva_plot(
  counts,
  geneset,
  method = "ssgsea",
  condition,
  title = "ssGSEA",
  compare = NULL
)
}
\arguments{
\item{counts}{Normalized count matrix - preferably z-score normalized by row. E.g. t(scale(t(counts(dds, normalized=T))))}

\item{geneset}{List with length > 1 of character vectors indicating gene sets of interest. Can be created using qusage::read.gmt("geneset.gmt")}

\item{method}{Method to employ in estimation of gene-set enrichment scores. One of "ssgsea","gsva","zscore","plage". Default is "ssgsea}

\item{condition}{Character vector of conditions in the same order they appear as columns in 'counts'.}

\item{title}{Character vector indicating the title of the plot}

\item{compare}{List of character vectors (each of length 2) indicating the pairwise comparisons to be made between conditions. Leave NULL for all comparisons to be made}
}
\value{
Violin plot of normalized enrichment scores for genesets between conditions
}
\description{
A function to run an ssGSEA analysis using normalized counts and create
a violin plot with pairwise comparisons
}
