% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmt.R
\name{read.gmt}
\alias{read.gmt}
\title{Read a gene set from file in gmt format}
\usage{
read.gmt(gmt)
}
\arguments{
\item{gmt}{character of path to file in gmt format}
}
\description{
This function read a gmt file to a list object in R
}
