% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zheat.R
\name{zheat}
\alias{zheat}
\title{Z-score normalized heatmap of gene expression}
\usage{
zheat(
  genes = NULL,
  counts,
  conditions,
  con = "WT",
  title = "DEGs",
  labgenes = NULL,
  zscore = T,
  rclus = F,
  hmcol = NULL,
  retClus = F,
  annoCols="Dark2"
)
}
\arguments{
\item{genes}{A character vector of genes to subset from rownames(counts) to create the heatmap}

\item{counts}{Normalized count matrix with rows as genes and columns as samples. Use 'counts(dds, normalized=T)'.}

\item{conditions}{Character vector indicating conditions belonging to each sample (same order as colnames(counts))}

\item{con}{Character vector indicating the control condition level in 'conditions'. Default is "WT".}

\item{title}{Character vector indicating the title of the figure}

\item{labgenes}{Character vector corresponding to rownames(counts) of genes to be labelled on the side of the heatmap. Leave as NULL to label all genes. Use "" to label no genes.}

\item{zscore}{Boolean indicating if counts should be z-score normalized (default=T)}

\item{rclus}{Boolean indicating if rows (genes) should be clustered. Leave FALSE for genes to be ordered on decreasing expression in 'con'.}

\item{hmcol}{Color ramp palette of length 100 for custom heatmap colouring. Leave NULL for defaults.}

\item{retClus}{Boolean indicating if clustered row names should be returnd. Default=FALSE.}
             
\item{annoCols}{Character vector indicating RColorBrewer palette name or colours for annotation colours if rclus is provided a data.frame.}
}
\value{
Heatmap of z-score normalized gene expression with control condition samples appearing first.
}
\description{
A function to generate a z-score normalized heatmap of gene expression from
a DESeq2 counts object.
}
